scripts/gera_slug_restaurante/requirements.txt
python-slugify==4.0.1
mysql-connector==2.2.9
